sdk-examples/python/calculate_api/pyproject.toml
[project]
name = "mcp_calculate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "nacos-sdk-python==2.0.4",
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.20",
]
